Extract some utils into misc
