+++ packages/ariakit-solid-core/src/group/group-label.tsx
@@ -5,7 +5,8 @@ import {
   onCleanup,
   useContext,
 } from "solid-js";
-import { createId, stableAccessor } from "../utils/reactivity.ts";
+import { createId } from "../utils/misc.ts";
+import { stableAccessor } from "../utils/reactivity.ts";
 import { createHook, createInstance } from "../utils/system.tsx";
 import type { Options, Props } from "../utils/types.ts";
 import { GroupLabelContext } from "./group-label-context.tsx";
