+++ packages/ariakit-solid-core/src/utils/misc.ts
@@ -1,0 +1,34 @@
+import { type MaybeAccessor, access } from "@solid-primitives/utils";
+import { type Accessor, type ValidComponent, createUniqueId } from "solid-js";
+
+/**
+ * Generates a unique ID.
+ */
+export function createId(
+  defaultId?: MaybeAccessor<string | undefined>,
+): Accessor<string> {
+  const id = createUniqueId();
+  return () => access(defaultId) ?? id;
+}
+
+/**
+ * Returns the tag name by parsing an element.
+ * @example
+ * function Component(props) {
+ *   const [ref, setRef] = createSignal();
+ *   const tagName = extractTagName(ref, "button"); // () => "div"
+ *   return <div ref={setRef} {...props} />;
+ * }
+ */
+export function extractTagName(
+  refOrElement?: MaybeAccessor<HTMLElement | undefined>,
+  fallback?: ValidComponent,
+) {
+  return () => {
+    const element = access(refOrElement);
+    return (
+      element?.tagName.toLowerCase() ??
+      (typeof fallback === "string" ? fallback : undefined)
+    );
+  };
+}
